Create Iron Box Using IOT
This project is aim to developing a frequent problem that our clothes get burnt while ironing. This mainly happens when the iron box is left idle when attending any phone calls or engaging in any other activities and also to save the power consumption. Mainly iron box switch ON through Wi-Fi connection, when Wi-Fi is connected to iron box then the circuit will ON. Here we used buzzer for indication when iron box get OF condition the buzzer gives beep indication to the user. 
                                                     The Smart Iron detects the position of the iron and allows the user to set an idle time at different positions. The circuit will switch OFF the iron box when it is not in use (no motion),in the rest position.